Remove all conversion from <A> to <B> warnings
This removes warnings for:
"'=': conversion from 'int' to 'float', possible loss of data"
"'=': conversion from 'float' to 'int', possible loss of data"

+++ desktop_version/src/Graphics.h
@@ -260,7 +260,7 @@ public:
 	std::vector <int> fadebars;
 
 	bool trinketcolset;
-	int trinketr, trinketg, trinketb;
+	Uint8 trinketr, trinketg, trinketb;
 
 	std::vector <textboxclass> textbox;
 	int ntextbox;
